apps/api/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-multipart
google-genai
python-dotenv
